requirements.txt
mysql-connector-python==9.0.0
passlib[bcrypt]==1.7.4
matplotlib==3.9.0
pillow==10.3.0
